Ability to set a default keyboard on startup
Rather than manually pressing a key each time you start the script to
identify the keyboard, this will attempt to look for a config.txt file,
containing a substring belonging to the keyboard device instance path.

If the config file contains a value, it will automatically set the
keyboard to the one which contains the substring. I'm using the vendor
ID, however you can use any part of the device instance path.

If no config file is found, then it will prompt you to press a key, and
then print out all detected keyboards. On the line labelled MACROS, you
can copy the VID or PID into a new config.txt file to remember that
device.