updated most scripts to be compatible with CC2017
older scripts for CC2015.4 can still be found in a branch that is
appropriately labeled. IDK exactly how it works!
Also, I added a feature to use the + and - keys on the numpad to change
the size of the program monitor from anywhere in the program.
Still to come:
must update premiere shortcuts .kys file
must update taran startup folder image
and mayeb some others.
also maybe should add a picture of the current state of my keyboards.