Massive improvements all around!
Many functions greatly improved in stablility, functionality, and
cross-compatibility.

This is the final update where I will be using the Logitech G15. I am
about to switch to the Corsair K95 RGB (with cherry mx browns!) which is
possibly the greatest keyboard in the world.

This update will be linked to from my upcoming 4-hour premiere tutorial.

Note that premiere keys sept 2016 FOR VISUALIZER.txt has not been
updated for awhile, and is probably not super accurate tot he actual key
commands anymore. Too lazy to update it right now.