MASSIVE RESTRUCTURING and other improvements
HOLY CRAP U GUYS
So in order to be able to launch INDUVIDUAL SCRIPTS using the new Stream
Deck, I had to learn how to use the #include feature. And later I had to
learn abotu the Sendmessage command in order to be able to send info
from one AHK script to abnother, to get the GUI working for those AHK
scripts...

Now, the actual FUNCTIONS are kept in a sperate script from the KEYBOARD
ASSIGNEMENTS that actually use them.

Also I made a lot of little improvements to script, but I don't rememebr
what.