Updating with my  changes
Displays all the keyboard and mouse button presses on the screen - Done
The Display should be customizable ie; changes could be made to the Font, Background Color, Font Color - Partially
The Display should decrease its transparency when the pressed key is released. - not done
The Display should disappear after a while if no further keys are pressed.- Done
Should display "234.56" as a whole group like "234.56" and not individual keystrokes like "2" then "3" then "4" then "." then "5" then "6". -Done
Same is applicable to words "HELLO" not "H" "E" "L" "L" "O" - Done
Should display "Numpad0" as "0" "Numpad1" as "1" etc. - Done 
Should display "(" as "(" and not "Shift+0" or "Shift+)" likewise for all symbols. - Done
Multiple keypresses like "Backspace Backspace Backspace Backspace Backspace" as "Backspace(x5)", the number in the brackets is to increase with the increase in keypresses. - Done
Should not treat "Hello" as "Hel(x2)o" - Done
Should refresh the display after a certain number of keypresses are made. - not done
For example, with a breakpoint of 15 characters the typed string "I went to the zoo yesterday had lots of fun" should be displayed as,
"I went to the z" then
"oo yesterday ha" and finally
"d lots of fun" - Done